# optimize BlockingCollection<T>.GetConsumingEnumerable iterator