Fixes #69320

I tried to search why this double-check had been done. It exists since the first RC of .NETCore 1.0. Could not find what version of .NET Framework introduced this change (just found that it exists in .NET 4.8)