@theodorzoulias 2022-05-21
Hi @pedrobsaila! Please take a look at [this](https://github.com/dotnet/runtime/issues/69320#issuecomment-1133587458) recent comment, regarding the behavior of the `BlockingCollection<T>` in an edge case. Your PR as it stands now probably changes the current behavior of the collection.