@pedrobsaila 2022-05-21
> I'm not sure if we could call this change an "optimization" unless there are demonstrable performance improvements in benchmarks.

Here's the source code of the benchmark https://github.com/dotnet/performance/commit/eb01ff12c7d35f6d2148df45a7d0d1dfbebb67dc

Here are the results :

* With double check :
```
AddRemoveFromSameThreads<Int32>.BlockingCollection: Job-SBTFJH(PowerPlanMode=00000000-0000-0000-0000-000000000000, InvocationCount=1, IterationTime=250.0000 ms, MaxIterationCount=20, MaxWarmupIterationCount=10, MinIterationCount=15, MinWarmupIterationCount=6, UnrollFactor=1, WarmupCount=-1) [Size=2000000]
BenchmarkDotNet=v0.13.1.1786-nightly, OS=Windows 10 (10.0.19044.1706/21H2/November2021Update)
Intel Core i7-10875H CPU 2.30GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK=7.0.100-preview.1.22110.4
  [Host]     : .NET 7.0.0 (7.0.22.7608), X64 RyuJIT
  Job-SBTFJH : .NET 7.0.0 (7.0.22.7608), X64 RyuJIT

PowerPlanMode=00000000-0000-0000-0000-000000000000  InvocationCount=1  IterationTime=250.0000 ms
MaxIterationCount=20  MaxWarmupIterationCount=10  MinIterationCount=15
MinWarmupIterationCount=6  UnrollFactor=1  WarmupCount=-1

|             Method |    Size |     Mean |    Error |   StdDev |   Median |      Min |      Max |      Gen 0 |     Gen 1 |     Gen 2 | Allocated |
|------------------- |-------- |---------:|---------:|---------:|---------:|---------:|---------:|-----------:|----------:|----------:|----------:|
| BlockingCollection | 2000000 | 495.6 ms | 27.30 ms | 31.44 ms | 494.6 ms | 435.3 ms | 550.0 ms | 31000.0000 | 2000.0000 | 2000.0000 | 276.15 MB |

AddRemoveFromSameThreads<String>.BlockingCollection: Job-SBTFJH(PowerPlanMode=00000000-0000-0000-0000-000000000000, InvocationCount=1, IterationTime=250.0000 ms, MaxIterationCount=20, MaxWarmupIterationCount=10, MinIterationCount=15, MinWarmupIterationCount=6, UnrollFactor=1, WarmupCount=-1) [Size=2000000]

// * Summary *

BenchmarkDotNet=v0.13.1.1786-nightly, OS=Windows 10 (10.0.19044.1706/21H2/November2021Update)
Intel Core i7-10875H CPU 2.30GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK=7.0.100-preview.1.22110.4
  [Host]     : .NET 7.0.0 (7.0.22.7608), X64 RyuJIT
  Job-SBTFJH : .NET 7.0.0 (7.0.22.7608), X64 RyuJIT

PowerPlanMode=00000000-0000-0000-0000-000000000000  InvocationCount=1  IterationTime=250.0000 ms
MaxIterationCount=20  MaxWarmupIterationCount=10  MinIterationCount=15
MinWarmupIterationCount=6  UnrollFactor=1  WarmupCount=-1

|             Method |    Size |     Mean |    Error |   StdDev |   Median |      Min |      Max |      Gen 0 |     Gen 1 |     Gen 2 | Allocated |
|------------------- |-------- |---------:|---------:|---------:|---------:|---------:|---------:|-----------:|----------:|----------:|----------:|
| BlockingCollection | 2000000 | 540.4 ms | 39.61 ms | 45.61 ms | 556.0 ms | 444.8 ms | 603.8 ms | 31000.0000 | 2000.0000 | 2000.0000 | 308.15 MB |
```
* Without double check :
```
AddRemoveFromDifferentThreads<Int32>.BlockingCollection: Job-YQTHKV(PowerPlanMode=00000000-0000-0000-0000-000000000000, Toolchain=CoreRun, InvocationCount=1, IterationTime=250.0000 ms, MaxIterationCount=20, MaxWarmupIterationCount=10, MinIterationCount=15, MinWarmupIterationCount=6, UnrollFactor=1, WarmupCount=-1) [Size=2000000]

BenchmarkDotNet=v0.13.1.1786-nightly, OS=Windows 10 (10.0.19044.1706/21H2/November2021Update)
Intel Core i7-10875H CPU 2.30GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK=7.0.100-preview.1.22110.4
  [Host]     : .NET 7.0.0 (7.0.22.7608), X64 RyuJIT
  Job-YQTHKV : .NET 7.0.0 (42.42.42.42424), X64 RyuJIT

PowerPlanMode=00000000-0000-0000-0000-000000000000  Toolchain=CoreRun  InvocationCount=1
IterationTime=250.0000 ms  MaxIterationCount=20  MaxWarmupIterationCount=10
MinIterationCount=15  MinWarmupIterationCount=6  UnrollFactor=1
WarmupCount=-1

|             Method |    Size |     Mean |    Error |   StdDev |   Median |      Min |      Max |      Gen 0 |     Gen 1 |     Gen 2 | Allocated |
|------------------- |-------- |---------:|---------:|---------:|---------:|---------:|---------:|-----------:|----------:|----------:|----------:|
| BlockingCollection | 2000000 | 387.7 ms | 11.85 ms | 13.65 ms | 389.0 ms | 362.5 ms | 417.2 ms | 30000.0000 | 1000.0000 | 1000.0000 | 260.14 MB |

AddRemoveFromDifferentThreads<String>.BlockingCollection: Job-YQTHKV(PowerPlanMode=00000000-0000-0000-0000-000000000000, Toolchain=CoreRun, InvocationCount=1, IterationTime=250.0000 ms, MaxIterationCount=20, MaxWarmupIterationCount=10, MinIterationCount=15, MinWarmupIterationCount=6, UnrollFactor=1, WarmupCount=-1) [Size=2000000]

BenchmarkDotNet=v0.13.1.1786-nightly, OS=Windows 10 (10.0.19044.1706/21H2/November2021Update)
Intel Core i7-10875H CPU 2.30GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK=7.0.100-preview.1.22110.4
  [Host]     : .NET 7.0.0 (7.0.22.7608), X64 RyuJIT
  Job-YQTHKV : .NET 7.0.0 (42.42.42.42424), X64 RyuJIT

PowerPlanMode=00000000-0000-0000-0000-000000000000  Toolchain=CoreRun  InvocationCount=1
IterationTime=250.0000 ms  MaxIterationCount=20  MaxWarmupIterationCount=10
MinIterationCount=15  MinWarmupIterationCount=6  UnrollFactor=1
WarmupCount=-1

|             Method |    Size |     Mean |   Error |   StdDev |   Median |      Min |      Max |      Gen 0 |     Gen 1 |     Gen 2 | Allocated |
|------------------- |-------- |---------:|--------:|---------:|---------:|---------:|---------:|-----------:|----------:|----------:|----------:|
| BlockingCollection | 2000000 | 403.4 ms | 9.33 ms | 10.75 ms | 401.6 ms | 386.2 ms | 424.8 ms | 31000.0000 | 2000.0000 | 2000.0000 | 276.14 MB |

```