@eiriktsarpalis 2022-06-08
> Without double check

Can you clarify what is meant by a "double check" in this case? Judging by the diff it seems that you have moved the check out of `TryTakeWithNoTimeValidation` and into the methods calling it.