@eiriktsarpalis 2022-06-09
@pedrobsaila I get that, but I fail to see which code path avoids checking `IsCompleted` twice. At short glance the diff seems to simply be moving the check out of the `TryTakeWithNoTimeValidation` method and into the methods that call it.