@eiriktsarpalis 2022-06-10
I see, so it basically only concerns this callsite?

https://github.com/dotnet/runtime/blob/95e8cae134f6cb046e2967bdcb7d8a89ff79ddfe/src/libraries/System.Collections.Concurrent/src/System/Collections/Concurrent/BlockingCollection.cs#L1652-L1659

Not necessarily convinced shaving a few percentage points is worth the potential risk of accidental regression. If anything, BlockingCollection has been superseded by more modern patterns such as [System.Threading.Channels](https://devblogs.microsoft.com/dotnet/an-introduction-to-system-threading-channels/) so we should encourage users to adopt that instead.